+++ src/main/java/no/ntnu/rentalroulette/APIController.java
@@ -1,4 +1,4 @@
-package no.ntnu.hello;
+package no.ntnu.rentalroulette;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -7,9 +7,10 @@ import org.springframework.web.bind.annotation.RestController;
  * Simple example controller
  */
 @RestController
-public class HelloController {
+public class APIController {
     @GetMapping("/hello")
     public String greeting() {
         return "Hello, World";
     }
+
 }
