+++ src/main/java/no/ntnu/rentalroulette/BackendApplication.java
@@ -1,13 +1,13 @@
-package no.ntnu.hello;
+package no.ntnu.rentalroulette;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class HelloApplication {
+public class BackendApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(HelloApplication.class, args);
+		SpringApplication.run(BackendApplication.class, args);
 	}
 
 }
